+++ index.js
@@ -1,9 +1,10 @@
 import React from 'react';
 import ReactDOM from 'react-dom';
-import { Widget } from './index_for_react_app';
+import { Widget, toggleWidget, openWidget, closeWidget, showWidget, hideWidget, isOpen, isVisible } from './index_for_react_app';
 
 const plugin = {
   init: (args) => {
+
     ReactDOM.render(
       <Widget
         socketUrl={args.socketUrl}
@@ -20,13 +21,19 @@ const plugin = {
         badge={args.badge}
         params={args.params}
       />, document.querySelector(args.selector)
-
     );
   }
 };
 
 export {
   plugin as default,
-  Widget
+  Widget,
+  toggleWidget as toggle,
+  openWidget as open,
+  closeWidget as close,
+  showWidget as show,
+  hideWidget as hide,
+  isOpen, 
+  isVisible
 };