+++ src/store/reducers/behaviorReducer.js
@@ -2,12 +2,24 @@ import { Map } from 'immutable';
 import * as actionTypes from '../actions/actionTypes';
 
 export default function (inputFieldTextHint) {
-  const initialState = Map({ initialized: false, showChat: false, disabledInput: false, inputFieldTextHint });
+  const initialState = Map({ initialized: false, showWidget: true, showChat: false, disabledInput: false, inputFieldTextHint });
 
   return function reducer(state = initialState, action) {
     switch (action.type) {
+      case actionTypes.SHOW_WIDGET: {
+        return state.update('showWidget', showWidget => true);
+      }
+      case actionTypes.HIDE_WIDGET: {
+        return state.update('showWidget', showWidget => false);
+      }
       case actionTypes.TOGGLE_CHAT: {
         return state.update('showChat', showChat => !showChat);
+      }
+      case actionTypes.OPEN_CHAT: {
+        return state.update('showChat', showChat => true);
+      }
+      case actionTypes.CLOSE_CHAT: {
+        return state.update('showChat', showChat => false);
       }
       case actionTypes.TOGGLE_INPUT_DISABLED: {
         return state.update('disabledInput', disabledInput => !disabledInput);